TVG-102 Blueprint for auth endpoints
