+++ api_blueprints/__init__.py
@@ -1,3 +1,4 @@
+from api_blueprints.auth import auth_blueprint
 from api_blueprints.guide import guide_blueprint
 from api_blueprints.reminders import reminder_blueprint, reminders_blueprint
 from api_blueprints.search_items import search_items_blueprint
